[SofaEulerianFluid] Move everything from module into a plugin.
By modularizing sofa we will decrease compilation time and footprint.
