+++ SofaAdvanced/CMakeLists.txt
@@ -19,7 +19,6 @@ install(FILES "${CMAKE_BINARY_DIR}/include/SofaAdvanced/config.h" DESTINATION "i
 
 set(SOFAADVANCED_TARGETS
     SofaComponentAdvanced
-    SofaEulerianFluid
     SofaNonUniformFem
     SofaSphFluid
     SofaVolumetricData
