+++ SofaAdvanced/config.h.in
@@ -26,13 +26,6 @@
 
 #cmakedefine SOFA_HAVE_FLOWVR
 
-#ifdef SOFA_BUILD_EULERIAN_FLUID
-#  define SOFA_TARGET SofaEulerianFluid
-#  define SOFA_EULERIAN_FLUID_API SOFA_EXPORT_DYNAMIC_LIBRARY
-#else
-#  define SOFA_EULERIAN_FLUID_API SOFA_IMPORT_DYNAMIC_LIBRARY
-#endif
-
 #ifdef SOFA_BUILD_SPH_FLUID
 #  define SOFA_TARGET SofaSphFluid
 #  define SOFA_SPH_FLUID_API SOFA_EXPORT_DYNAMIC_LIBRARY
