+++ applications/plugins/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(../../extlibs/CImg extlibs/CImg)
 
 find_package(SofaFramework)
 
+sofa_add_plugin(SofaEulerianFluid SofaEulerianFluid)
 sofa_add_plugin(image image)
 sofa_add_plugin(Compliant Compliant)
 sofa_add_plugin(CGALPlugin CGALPlugin) # Depends on image
