+++ modules/SofaComponentAdvanced/CMakeLists.txt
@@ -11,7 +11,7 @@ set(SOURCE_FILES
 )
 
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} PUBLIC SofaEulerianFluid SofaNonUniformFem SofaSphFluid SofaVolumetricData)
+target_link_libraries(${PROJECT_NAME} PUBLIC SofaNonUniformFem SofaSphFluid SofaVolumetricData)
 set_target_properties(${PROJECT_NAME} PROPERTIES COMPILE_FLAGS "-DSOFA_BUILD_COMPONENT_ADVANCED")
 set_target_properties(${PROJECT_NAME} PROPERTIES PUBLIC_HEADER "${HEADER_FILES}")
